o/~ Signs, signs, everywhere signs. Fuckin' up the code, breakin' my mind

+++ vm/llvm/jit_util.cpp
@@ -276,11 +276,11 @@ extern "C" {
   }
 
   Object* rbx_construct_splat(STATE, Arguments& args, size_t start, size_t total) {
-    size_t splat_size = args.total() - total;
+    int splat_size = args.total() - total;
     if(splat_size > 0) {
       Array* ary = Array::create(state, splat_size);
 
-      for(size_t i = 0, n = start; i < splat_size; i++, n++) {
+      for(int i = 0, n = start; i < splat_size; i++, n++) {
         ary->set(state, i, args.get_argument(n));
       }
       return ary;
